refactor(pagination): refactor pagination component
